Add prettier
