+++ ngx-google-translate-ui/.prettierignore
@@ -1,1 +1,4 @@
 package.json
+
+/dist
+/coverage
