+++ ngx-google-translate-ui/.prettierrc
@@ -1,0 +1,7 @@
+{
+	"arrowParens": "avoid",
+	"printWidth": 80,
+	"proseWrap": "always",
+	"singleQuote": true,
+	"trailingComma": "none"
+}
